+++ Sources/BlueTriangle/Protocols/CrashReportManaging.swift
@@ -1,0 +1,12 @@
+//
+//  CrashReportManaging.swift
+//
+//  Created by Mathew Gacy on 4/11/22.
+//  Copyright © 2022 Blue Triangle. All rights reserved.
+//
+
+import Foundation
+
+protocol CrashReportManaging {
+    func uploadReports(session: Session)
+}
