+++ Sources/BlueTriangle/Protocols/FileLocation.swift
@@ -1,0 +1,12 @@
+//
+//  FileLocation.swift
+//
+//  Created by Mathew Gacy on 4/11/22.
+//  Copyright © 2022 Blue Triangle. All rights reserved.
+//
+
+import Foundation
+
+protocol FileLocation {
+    var containerURL: URL? { get }
+}
