+++ Sources/BlueTriangle/Protocols/Uploading.swift
@@ -1,0 +1,12 @@
+//
+//  Uploading.swift
+//
+//  Created by Mathew Gacy on 4/8/22.
+//  Copyright © 2022 Blue Triangle. All rights reserved.
+//
+
+import Foundation
+
+protocol Uploading {
+    func send(request: Request)
+}
